# server/resource/Test2_db/ProjectApproval.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE ProjectApproval.js PLEASE EDIT ../custom/ProjectApprovalCustom.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
const app = require('../../app.js');
const db_Test2_db = require('../../db/Test2_db_schema.js')
const properties = require('../../properties.js');
const handleError = require('../../security/util.js').handleError;
require('./custom/ProjectApprovalCustom.js');

/*
 * SCHEMA DB ProjectApproval
 * 
	{
		CreatedById: {
			type: 'Integer'
		},
		CreatedOn: {
			type: 'Date'
		},
		IsActive: {
			type: 'Boolean'
		},
		Level: {
			type: 'Integer'
		},
		ModifiedById: {
			type: 'Integer'
		},
		ModifiedOn: {
			type: 'Date'
		},
		ProjectId: {
			type: 'Integer'
		},
		UserId: {
			type: 'Integer'
		},
		UserTypeId: {
			type: 'Integer'
		},
		VendorId: {
			type: 'Integer'
		},
		//RELATIONS
		
		
		//EXTERNAL RELATIONS
		
		
	}
 * 
 */



//CRUD METHODS


/**
 * ProjectApprovalService.create
 *   @description CRUD ACTION create
 *
 */
app.post(properties.api + '/projectapprovals', function (req, res) {
	obj = new db_Test2_db.ProjectApproval(req.body);
	obj.save(function (err) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * ProjectApprovalService.delete
 *   @description CRUD ACTION delete
 *   @param ObjectId id Id
 *
 */
app['delete'](properties.api + '/projectapprovals/:id', function (req, res) {
	db_Test2_db.ProjectApproval.findByIdAndRemove(req.params.id, function (err) {
		  if (err) return handleError(err, res);
		  res.send(err);
	});
});

/**
 * ProjectApprovalService.get
 *   @description CRUD ACTION get
 *   @param ObjectId id Id 
 *
 */
app.get(properties.api + '/projectapprovals/:id', function (req, res) {
	db_Test2_db.ProjectApproval.findOne({_id:req.params.id}).exec(function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * ProjectApprovalService.list
 *   @description CRUD ACTION list
 *
 */
app.get(properties.api + '/projectapprovals', function (req, res) {
	db_Test2_db.ProjectApproval.find().exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
});

/**
 * ProjectApprovalService.update
 *   @description CRUD ACTION update
 *   @param ObjectId id Id
 *
 */
app.post(properties.api + '/projectapprovals/:id', function (req, res) {
	db_Test2_db.ProjectApproval.findByIdAndUpdate(req.params.id, req.body, {'new': true}, function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */